Wonjin_amount int형에서 float형으로 변경

+++ src/main/java/com/donation/common/request/post/PostSaveReqDto.java
@@ -20,13 +20,13 @@ public class PostSaveReqDto {
     private String content;
 
     @NotNull(message = "후원량이 없습니다.")
-    private Integer amount;
+    private float amount;
 
     @NotNull(message = "카테고리를 선택해주세요.")
     private Category category;
 
     @Builder
-    public PostSaveReqDto(String title, String content, Integer amount, Category category) {
+    public PostSaveReqDto(String title, String content, float amount, Category category) {
         this.title = title;
         this.content = content;
         this.amount = amount;
