+++ src/main/java/com/donation/common/response/post/PostFindRespDto.java
@@ -19,7 +19,7 @@ public class PostFindRespDto {
 
     private Write write;
 
-    private int amount;
+    private float amount;
 
     private Category category;
 
@@ -31,7 +31,7 @@ public class PostFindRespDto {
     private Integer favoriteCount;
 
     @QueryProjection
-    public PostFindRespDto(Long postId, Long userId, String username, String name, String profileImage, Write write, int amount, Category category, PostState state) {
+    public PostFindRespDto(Long postId, Long userId, String username, String name, String profileImage, Write write, float amount, Category category, PostState state) {
         this.postId = postId;
         this.userRespDto = new UserRespDto(userId, username, name, profileImage);
         this.write = write;
@@ -42,7 +42,7 @@ public class PostFindRespDto {
 
 
     @Builder
-    public PostFindRespDto(Long postId, UserRespDto userRespDto, Write write, int amount, Category category, PostState state, List<String> postDetailImages, Integer favoriteCount) {
+    public PostFindRespDto(Long postId, UserRespDto userRespDto, Write write, float amount, Category category, PostState state, List<String> postDetailImages, Integer favoriteCount) {
         this.postId = postId;
         this.userRespDto = userRespDto;
         this.write = write;
